Jest setup complete
First cookie test is complete
