+++ client/tsconfig.json
@@ -41,6 +41,9 @@
 		"src/**/*.json",
 		"src/**/*.svelte",
 		"mock/**/*.js",
-		"mock/**/*.ts"
+		"mock/**/*.ts",
+		"__tests__/**/*.js",
+		"__tests__/**/*.ts",
+		"__tests__/**/*.json"
 	]
 }
